requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
DateTime==4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-import-export==3.0.0b4
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
tqdm==4.64.1
tzdata==2022.6
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.5.1
